pyproject.toml
[project]
name = "graphrag-apollo11"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.42.15",
    "chromadb>=1.3.7",
    "graphrag>=2.7.0",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
]
